Fixed the offset for Forest Tiles and updated collision
The pivot point of the forest tile prefabs were outside of the tiles, causing the player and the exit to occasionally spawn outside of bounds.

Changed the dash to check for layers called tile layer to avoid dashing through collision. Additionally, changed the forest tiles to be under the 'tile layer' layer